Fixes small memory leak in SC.ObserverSet.
Each time that a new observer is added to an object the ObserverSet for the object will add a tracking hash for the target and the method.  As more methods are tracked for the target, they are added and as methods are no longer tracked for the target, they are removed.  However, even when no methods are tracked for the target, the tracking hash for the target still exists.  This creates a small leak of memory, because the target may have been destroyed and freed, but we are still maintaining an empty tracking hash for it.

Here's an example that would create 3,000 empty hashes in coreOb._kvo_observers_key._members:

    obs = [];
    window.coreOb = SC.Object.create({ key: '1' });
    SC.run(function() {
      for (var i = 2999; i >= 0; i--) {
        obs.push(SC.Object.create({ myKeyBinding: SC.Binding.from('coreOb.key') }));
      }
    });

    SC.run(function() {
      for (var i = obs.length - 1; i >= 0; i--) {
        obs.pop().destroy();
      }
    });