Makes invokeNext trigger the next run of the run loop if none is scheduled.   This makes invokeNext the appropriate method to use when needing the DOM to be up-to-date before running code against it.
Previously, invokeNext functions would not run until the next run loop occurred, which if there were no events or timers, might mean that the code won't run for a long time.  While invokeLast lets you put code at the end of the run loop, the DOM will not have yet been updated and invokeLater is risky to use at small intervals, because if the interval is too small, the invokeLater code will run in the current run of the run loop and again the DOM will not have been updated.  Instead, invokeNext now ensures that the run loop completes execution and schedules an immediate timeout in order to allow the browser to update the DOM.

This works with the timer additions to the run loop, so additional run loops only run once even if there is an invokeLater expiring at the same time as the invokeNext timeout triggers a new run loop.

- Includes unit test.
- Fixes bug in previous unit test that left a run loop unfinished.