Update button isDefault/isCancel handling on SC.AlertPane.
If any button specifies isDefault, ignore the default value of this
property on each button and always use the value passed in. This allows
just specifying isDefault on the one button that should be default,
without having to also say isDefault:NO on all the other buttons.

Also do the same for isCancel.